added authorization

+++ src/Components/BlogPage.jsx
@@ -1,14 +1,19 @@
 import { useDispatch, useSelector } from "react-redux";
 import { useNavigate, useParams } from "react-router-dom";
 import store from "../Utils/Store";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import { deleteBlog } from "../Utils/constants";
 import { removeBlogs, removeUserBlogs } from "../Utils/BlogSlice";
+import { Modal } from './Exports';
 
 export default function BlogPage () {
 
+  const [ showModal, setShowModal ] = useState( false );
+  const [ deleteblog, setDeleteblog ] = useState( false );
+
   const navigate = useNavigate();
   const dispatch = useDispatch();
+  
   const isLoggedIn = useSelector( store => store?.user?.isLoggedIn );
   const user = useSelector( store => store?.user?.User );
 
@@ -18,12 +23,12 @@ export default function BlogPage () {
     if ( blog?.blogId == id ) return blog;
   } );
 
-  const handleDelete = () => {
+  if ( deleteblog ) {
     const data = deleteBlog( id, user );
     dispatch( removeBlogs() );
     dispatch( removeUserBlogs() );
     navigate( '/' );
-  };
+  }
 
   useEffect( () => {
     if ( !blogs ) {
@@ -32,14 +37,16 @@ export default function BlogPage () {
   }, [] );
 
   return (
-    <div className='flex flex-col items-center pt-6 gap-10'>
+    <div className=' relative flex flex-col items-center pt-6 gap-10 font-ubuntu min-h-[85vh]'>
+      <Modal show={ showModal } Delete={ ()=>setDeleteblog(!deleteblog) } />
       <h1 className='font-semibold text-4xl my-6'>{ thisItemData && thisItemData[ 0 ]?.title }</h1>
+      <p className='text-lg max-w-[40%] min-w-[400px]'>{ thisItemData && thisItemData[ 0 ]?.description }</p>
       <p className='text-lg max-w-[40%] min-w-[400px]'>{ thisItemData && thisItemData[ 0 ]?.blog }</p>
       <p className='text-xl font-semibold'>~ { thisItemData && thisItemData[ 0 ]?.username }</p>
       { isLoggedIn && user == thisItemData[ 0 ]?.username && id == thisItemData[ 0 ]?.blogId ?
         <button
-          onClick={ handleDelete }
-          className="bg-red-600 py-3 px-5 font-medium my-2">
+          onClick={ () => setShowModal( !showModal ) }
+          className="bg-red-600 py-3 px-5 font-medium my-2 text-white rounded-full">
           Delete
         </button> : '' }
     </div>
